Save some BtreeMap lookups
Previously, code would do more lookups into the BtreeMap datastructure,
together with an internal traversal of it, depending on whether the
key is present in the map or not.

If it is present, it would do two lookups, now it only does one.
If it isn't present, it would do three lookups, now it only does two.